+++ snakemake/linting/rules.py
@@ -48,7 +48,7 @@ class RuleLinter(Linter):
     def lint_not_used_params(
         self,
         rule,
-        valid_names={"input", "output", "log", "params"},
+        valid_names={"input", "output", "log", "params", "wildcards", "threads"},
         regex=re.compile("{{(?P<name>{}).*?}}".format(NAME_PATTERN)),
     ):
         if rule.shellcmd:
@@ -56,7 +56,9 @@ class RuleLinter(Linter):
                 name = match.group("name")
                 if name not in valid_names:
                     yield Lint(
-                        title="Shell command directly uses variables from outside of the rule",
+                        title="Shell command directly uses variable {} from outside of the rule".format(
+                            name
+                        ),
                         body="It is recommended to pass all files as input and output, and non-file parameters "
                         "via the params directive. Otherwise, provenance tracking is less accurate.",
                         links=[links.params],